requirements.txt
fastapi
uvicorn
sqlmodel
phonenumbers
pydantic[email]
python-dotenv
python-jose
passlib
bcrypt